Moved general-purpose functions and classes to a new utils directory; added docstrings
